requirements.txt
# Yanshee Model Server Dependencies

# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0
matplotlib>=3.7.0

# Web framework
Flask>=2.3.0
flask-cors>=4.0.0

# AI/ML models
transformers>=4.35.0
depth-anything-3>=0.1.0
sam3>=0.1.0

# ROS API (local dependency)
# ros_api is a local module, not available on PyPI

# Development dependencies (optional)
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
pytest>=7.4.0

# Note: Some models may require specific versions or additional dependencies
# Check model documentation for specific requirements
